Add a mechanism to disable dimension tick tags
This is used in a similar manner to the mechanism that disables the
forceloaded chunks. Simply delete the function
'lantern_impl:v1/enable_dimension_tags', and if you wish to indicate
that you want the dimension tags anyway, use the fake player
'lantern.dimension_tag lantern.counts'.

This commit also moves the commands that temporarily disable the
dimension tick tags during load to not rely on forceloading being
enabled; this is so that if forceloading was enabled on a previous load,
the command blocks will not keep running for no reason.