Use forceload command directly
It has come to light that 1.14.3 allows functions to run 'forceload', so
to guarantee instant loading and avoid pestering the user, Lantern now
directly forceloads the chunks it requires.

In accordance with this change, the ability to disable forceloading has
been removed, as there is no longer any downside for the end user or
other data packs. The ability to disable dimension tags still exists.

This commit also makes a few minor changes, mostly related to internal
organization. However, an important change is that the shulker box in
the forceloaded chunks is now yellow instead of purple. It would have
been disasterous if the purple shulker box remained in Lantern 1.0, as
Lantern only provides the content dropping loot table for yellow shulker
boxes.