Alter the dimension tag option's implementation
The dimension tag option was previously implemented using a fake player
that tracked the 'lantern.counts' objective, however this approach is
misleading, as it implies the number of times the dimension tag option
is enabled actually matters. Therefore, the option now uses a fake
player that tracks the 'lantern.global' objective.