Create flags system
As the previous change to the dimension tag option backfired (due to
'lantern.global' not resetting on reload), a new 'lantern.flags'
scoreboard objective has been created. This objective exists to keep
ephemeral configuration options out of the 'lantern.counts' objective,
while not requiring the data pack creator to reset the fake player
manually each enumeration.

In accordance with this change, the dimension tag option is now
considered a "flag" and uses the 'lantern.flags' scoreboard objective.