Simply version resolution process
The two-phase version resolution process works for medium-complexity
library modules, but it's a bit much for low-complexity modules and
insufficient for high-complexity modules.

As it turns out, despite being fairly lightweight, Lantern's load
process is high-complexity due to the flags system. The fact that the
core library providing the enumerate / resolve pattern could not itself
load solely with that pattern made it clear that the process should be
simplified such that modules may implement the patterns they require
on their own.

Therefore, the only public-facing tag is now '#lantern.1:resolve'. The
tag-ordering guarantees added with 1.14.1 ensure that dependencies load
before dependents, so long as they are bundled together. Dependencies
are free to implement their own enumerate / resolve pattern, as Lantern
has, but the majority of data packs should be alright with just resolve.

This commit also removes the 'lantern.loaded' scoreboard, as it is
largely redundant when data packs are given greater control over the
version resolution process. A simple module can wait until a successful
resolution before settings its 'lantern.versions' value, while a complex
module may need to set the value before successful resolution, only to
reset it in the case of failure.