Cleanup for 1.0
This commit adds a 'post_resolve' tag for other data packs to use if
they wish to create a flags system similar to that used by Lantern
itself.

The vestigial 'marker' entity type tag has been removed from Lantern,
after the cleanup code was removed in favor of data packs handling their
own cleanup (originally with 'schedule', now with 'post_resolve').

Debug 'say' commands in the 'initialize' function were removed.

Dimension tags now run at '-30000000 0 8880' to emphasize that the
position of the command block should not be relied on.

The 'pack.mcmeta' file was removed from Lantern to emphasize that the
library is made to be bundled with other data packs. An unbundled
version of Lantern will be made available for testing purposes each
release, but the README and documentation will make it clear that
Lantern must be bundled before distribution.