Lantern 1.1.0
Add the 'forceloading' flag, controlled by the presence of the
'lantern.1:flags/enable_forceloading' function. This flag controls
whether Lantern forceloads and initializes the chunk at
'-30000000 0 8884'. If Lantern previously ran with dimension tags
enabled, it will temporarily forceload the chunk to ensure that the
dimension tags do not run. However, Lantern will not modify blocks in
the chunk, and will return the chunk to being un-forceloaded if it was
previously in that state.

Clean up old dimension tag command blocks in a more precise manner. For
1.0.0 and 1.0.1 command blocks, the 'Command' field is used to target
command blocks for deletion. For 1.1.0+ command blocks, the 'CustomName'
field is used instead, as the 'Command' field changes too frequently for
its continued use to be sustainable.

Add a second yellow shulker box at '-30000000 0 8883' in each dimension.
Slots 1-26 in this shulker box must remain empty at all times. Slot 0
may contain an item. This shulker box is useful for computational loot
tables that return data in a single item's NBT; the slot 0 restriction
avoids contextually useless serializations during this process.

Add a dropper at '-30000000 0 8884' in each dimension. Data packs do not
modify its contents; rather, they create a loot table that drops a
certain number of bows depending on which conditions were met, and then
use 'loot insert -30000000 0 8884'. Due to a quirk that Lantern has
exploited, this will successfully return the number of bows the loot
table dropped, despite the dropper being completely full.

Populate the tag 'Book.tag.lantern.Dimension' in each dimension's
forceloaded lectern, with a different 'int' value depending on the
dimension of the lectern. The value for the Overworld is 0, the value
for The Nether is -1, and the value for The End is 1.